requirements.txt
tqdm
imgaug
pandas==0.25.3
tensorflow==2.6.0
tensorflow-serving-api==2.6.0
keras==2.6.0
tensorflow-estimator==2.6.0
requests
scipy==1.4.1
pylint==2.4.3
Pillow
numpy==1.19.2
scikit-learn==0.21.1
imageio
flask==1.1.1
lxml==4.5.2
prometheus-client==0.7.1
pytesseract==0.3.4
joblib
Werkzeug==0.16.0
python-levenshtein==0.12.1
colormath
google-cloud-vision==1.0.0
Unidecode==1.1.1
pyenchant==3.1.1
torch==1.11.0
tensorflow-hub==0.8.0
python-consul==1.1.0
lightgbm==3.3.1
opencv-contrib-python==4.5.1.48
opencv-python-headless==4.5.1.48
opencv-python==4.5.1.48
itsdangerous==2.0.1
jinja2==3.0.3
